requirements.txt
python-dotenv
discogs_client